pyproject.toml
[tool.poetry]
name = "opendu"
version = "0.0.1"
description = "OpenDu, OpenCUI Language Understanding and Generation"
authors = ["Xiaoyun Wu <xiaoyun.wu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.9.13"
sentence-transformers = "^2.2.2"
langchain = ">=0.0.315,<0.1.0"
pybars3 = ">=0.9.7,<0.10.0"
Jinja2 = "^3.1.5"
datasets = ">=2.14.4,<2.15.0"
pydantic = ">=2.4.2,<2.5.0"
transformers = "==4.35.*"
numpy = ">=1.26.1,<1.27.0"
aiohttp = ">=3.8.6,<3.9.0"
pyyaml = ">=6.0.1,<6.1.0"
pandas = ">=2.1.1,<2.2.0"
evaluate = ">=0.4.0,<0.5.0"
scikit-learn = ">=1.3.0,<1.4.0"
peft = ">=0.5.0,<0.6.0"
torch = ">=2.0.1,<2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
